backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
cryptography>=41.0.0
python-multipart>=0.0.6
email-validator>=2.1.0
aiosqlite>=0.19.0
httpx>=0.26.0
